Refactor and pass build
