+++ src/core/wallets/evm_wallet/index.ts
@@ -1,1 +1,2 @@
 export { EdwinEVMWallet } from './evm_wallet';
+export { EdwinEVMPublicKeyWallet } from './evm_public_key_wallet';
