+++ src/plugins/uniswap/uniswapPlugin.ts
@@ -10,6 +10,19 @@ export class UniswapPlugin extends EdwinPlugin {
     }
 
     getTools(): Record<string, EdwinTool> {
+        // Combine public and private tools
+        return {
+            ...this.getPublicTools(),
+            ...this.getPrivateTools(),
+        };
+    }
+
+    getPublicTools(): Record<string, EdwinTool> {
+        // Uniswap has no public tools
+        return {};
+    }
+
+    getPrivateTools(): Record<string, EdwinTool> {
         const uniswapProtocol = this.toolProviders.find(
             provider => provider instanceof UniswapProtocol
         ) as UniswapProtocol;
